requirements.txt
celery==5.4.0
Django==4.2.18
django-celery-beat==2.7.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
factory_boy==3.3.1
Faker==33.3.1
python-dotenv==1.0.1
